Fixed being unable to extract video album and title
When there was no artist metadata on a video all other metadata
was also ignored. This is fixed for new imports. For old imports
please run the "Rescan video files for missing metadata" task.
This will add missing metadata from the video files to the database.
It will not clobber any existing metadata in the DB.