Render gallery slideshow previews on mobile
Slideshow previews were only triggered on mousenter and mousexit.
On mobile they're now activated if the item is fully in view.
Additionally there's loading changes to previews. They now won't
load until the thumbnail has and if a slideshow is meant to be
active they will now trigger on load if they weren't.